fix: retry pg connection on new library code
